se agrega logica para que cuando se crea la orden de ingreso aumente la cantidad del inventario o cree un nuevo inventario si el producto no esta

+++ src/main/java/com/juan/curso/springboot/webapp/gestordedepositos/Dtos/InventarioDTO.java
@@ -18,7 +18,7 @@ public class InventarioDTO {
     private Long id_inventario;
     private Producto producto;
     private Ubicacion ubicacion;
-    private Double cantidad;
+    private int cantidad;
     private Date fecha_actualizacion;
 
     public InventarioDTO(Inventario inventario) {
